move `groupby_impltypeflags` from `codegenerator` to `typedesc`.
